% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author_names.R
\docType{data}
\name{author_names}
\alias{author_names}
\title{NBER working paper author names}
\format{
Data frame with columns
\describe{
\item{author}{Author ID.}
\item{name}{Author name.}
\item{user_nber}{Author username on NBER website.}
\item{user_repec}{Author username on RePEc.}
}
}
\source{
\href{https://data.nber.org/nber-wp-logs/}{National Bureau of Economic Research}
}
\usage{
data(author_names)
}
\description{
Data frame containing author names.
}
\examples{
author_names

if (require('dplyr')) {
paper_authors \%>\% count(author) \%>\% left_join(author_names)
}

}
\keyword{datasets}
