% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_programs.R
\docType{data}
\name{paper_programs}
\alias{paper_programs}
\title{NBER working paper programs}
\format{
Data frame with columns
\describe{
\item{paper}{Working paper number}
\item{program}{Program code}
}
}
\source{
\href{https://data.nber.org/nber-wp-logs/}{National Bureau of Economic Research}
}
\usage{
data(paper_programs)
}
\description{
Crosswalk between NBER working papers and programs.
}
\examples{
paper_programs

if (require('dplyr')) {
paper_programs \%>\% count(program) \%>\% left_join(programs)
}

}
\keyword{datasets}
