% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{papers}
\alias{papers}
\title{NBER working papers}
\format{Data frame with columns
\describe{
\item{number}{Number}
\item{year}{Publication year}
\item{month}{Publication month}
\item{title}{Title}
}}
\source{
\href{https://www.nber.org/RePEc/nbr/nberwo/}{National Bureau of Economic Research}
}
\usage{
data(papers)
}
\description{
Catalogue of NBER working papers published between 1973 and 2019.
}
\keyword{datasets}
