% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program_descriptions.R
\docType{data}
\name{program_descriptions}
\alias{program_descriptions}
\title{NBER working paper program descriptions}
\format{
Data frame with columns
\describe{
\item{program}{Program code}
\item{program_desc}{Program description}
}
}
\source{
\href{https://www.nber.org/RePEc/nbr/nberwo/}{National Bureau of Economic Research}
}
\usage{
data(program_descriptions)
}
\description{
Crosswalk between NBER program codes and descriptions.
}
\keyword{datasets}
