% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\docType{data}
\name{programs}
\alias{programs}
\title{NBER working paper programs}
\format{
Data frame with columns
\describe{
\item{paper}{Working paper number}
\item{program}{Program code}
}
}
\source{
\href{https://www.nber.org/RePEc/nbr/nberwo/}{National Bureau of Economic Research}
}
\usage{
data(programs)
}
\description{
Crosswalk between NBER working papers and programs.
}
\keyword{datasets}
