% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R
\docType{data}
\name{authors}
\alias{authors}
\title{NBER working paper authors}
\format{
Data frame with columns
\describe{
\item{author}{Author ID}
\item{name}{Author name}
\item{user_nber}{Author username on NBER website}
\item{user_repec}{Author username on RePEc}
}
}
\source{
\href{https://data.nber.org/nber-wp-logs/}{National Bureau of Economic Research}
}
\usage{
data(authors)
}
\description{
Data frame containing author attributes.
}
\examples{
authors

if (require('dplyr')) {
paper_authors \%>\% count(author) \%>\% left_join(authors)
}

}
\keyword{datasets}
