% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_authors.R
\docType{data}
\name{paper_authors}
\alias{paper_authors}
\title{NBER working paper authors}
\format{
Data frame with columns
\describe{
\item{paper}{Working paper number.}
\item{author}{Author ID.}
}
}
\source{
\href{https://data.nber.org/nber-wp-logs/}{National Bureau of Economic Research}
}
\usage{
data(paper_authors)
}
\description{
Data frame containing paper-author correspondences.
}
\examples{
paper_authors

if (require('dplyr')) {
paper_authors \%>\% count(author) \%>\% left_join(authors)
}

}
\keyword{datasets}
