% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program_descriptions.R
\docType{data}
\name{program_descriptions}
\alias{program_descriptions}
\title{NBER working paper program descriptions}
\format{
Data frame with columns
\describe{
\item{program}{Program code}
\item{program_desc}{Program description}
\item{program_category}{Program category based on Chari and Goldsmith-Pinkham (2017)}
}
}
\source{
\href{https://www.nber.org/RePEc/nbr/nberwo/}{National Bureau of Economic Research}
}
\usage{
data(program_descriptions)
}
\description{
Crosswalk between NBER program codes, descriptions and categories.
}
\references{
Chari, A. and P. Goldsmith-Pinkham (2017).
Gender Representation in Economics Across Topics and Time: Evidence from the NBER Summer Institute.
NBER Working Paper No. 23953,
National Bureau of Economic Research.
}
\keyword{datasets}
